+++ packages/components/accordion/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2](https://github.com/adevinta/spark/compare/@spark-ui/accordion@0.4.1...@spark-ui/accordion@0.4.2) (2024-06-20)
+
+### Bug Fixes
+
+- **accordion:** updated accordion header height ([cd02a17](https://github.com/adevinta/spark/commit/cd02a1770b151fff05000a8f6f6cb17f8bdf5767))
+
+## [0.4.1](https://github.com/adevinta/spark/compare/@spark-ui/accordion@0.4.0...@spark-ui/accordion@0.4.1) (2024-06-19)
+
+**Note:** Version bump only for package @spark-ui/accordion
+
+# [0.4.0](https://github.com/adevinta/spark/compare/@spark-ui/accordion@0.3.0...@spark-ui/accordion@0.4.0) (2024-06-18)
+
+### Features
+
+- **accordion:** outlined and filled designs ([0968a8d](https://github.com/adevinta/spark/commit/0968a8dcadfbdeceb64459d0b9446670b88d51fd))
+
+# [0.3.0](https://github.com/adevinta/spark/compare/@spark-ui/accordion@0.2.0...@spark-ui/accordion@0.3.0) (2024-06-17)
+
+### Features
+
+- **accordion:** accordion item header ([bef11fc](https://github.com/adevinta/spark/commit/bef11fcc89f4d8492883ef276e2c6ee780dbfc09))
+
+# [0.2.0](https://github.com/adevinta/spark/compare/@spark-ui/accordion@0.1.1...@spark-ui/accordion@0.2.0) (2024-06-17)
+
+### Features
+
+- **accordion:** add dependencies ([40a10ad](https://github.com/adevinta/spark/commit/40a10adf482e0fb5ae2eab96aa793177a29c2a85))
+- **accordion:** composition with collapsible ([7dd8446](https://github.com/adevinta/spark/commit/7dd8446d40f2a1331327ba3c4c70ba9ff85c584e))
+- **accordion:** split props ([309ec9b](https://github.com/adevinta/spark/commit/309ec9bf356aaf8875e03b0929e815891666da50))
+
 ## [0.1.1](https://github.com/adevinta/spark/compare/@spark-ui/accordion@0.1.0...@spark-ui/accordion@0.1.1) (2024-06-15)
 
 **Note:** Version bump only for package @spark-ui/accordion